remove fixtures abstraction
