--- src/fixtures/user.ts
@@ -1,31 +1,0 @@
-import { UserProfile } from "../types";
-
-export const user: UserProfile = {
-  userId: "abc123",
-  displayName: "Jane Smith",
-  webUrl: "",
-  apiUrl: "",
-  avatar: "",
-  secureAvatarUrl: "",
-  badge: [],
-  privateFields: {
-    canPostComment: true,
-    isPremoderated: false,
-    hasCommented: true
-  }
-};
-
-export const staffUser: UserProfile = {
-  userId: "abc123",
-  displayName: "Jane Smith",
-  webUrl: "",
-  apiUrl: "",
-  avatar: "",
-  secureAvatarUrl: "",
-  badge: [{ name: "Staff" }],
-  privateFields: {
-    canPostComment: true,
-    isPremoderated: false,
-    hasCommented: true
-  }
-};
